move data transform functions to shared folder
