+++ src/app/routes/idx/getInitialData/index.js
@@ -1,10 +1,10 @@
 import pipe from 'ramda/src/pipe';
 import fetchPageData from '../../utils/fetchPageData';
-import filterUnknownContentTypes from '../../utils/filterUnknownContentTypes';
-import filterEmptyGroupItems from '../../utils/filterEmptyGroupItems';
-import squashTopStories from '../../utils/squashTopStories';
-import addIdsToItems from '../../utils/addIdsToItems';
-import filterGroupsWithoutStraplines from '../../utils/filterGroupsWithoutStraplines';
+import filterUnknownContentTypes from '#app/routes/utils/sharedDataTransformers/filterUnknownContentTypes';
+import filterEmptyGroupItems from '#app/routes/utils/sharedDataTransformers/filterEmptyGroupItems';
+import squashTopStories from '#app/routes/utils/sharedDataTransformers/squashTopStories';
+import addIdsToItems from '#app/routes/utils/sharedDataTransformers/addIdsToItems';
+import filterGroupsWithoutStraplines from '#app/routes/utils/sharedDataTransformers/filterGroupsWithoutStraplines';
 
 const transformJson = pipe(
   filterUnknownContentTypes,
