+++ src/app/routes/utils/sharedDataTransformers/filterEmptyGroupItems/index.js
