+++ src/app/routes/utils/sharedDataTransformers/filterEmptyGroupItems/index.test.js
