+++ src/app/routes/utils/sharedDataTransformers/filterGroupsWithoutStraplines/index.js
