+++ src/app/routes/utils/sharedDataTransformers/filterGroupsWithoutStraplines/index.test.js
