+++ src/app/routes/utils/sharedDataTransformers/squashTopStories/index.js
