+++ src/app/routes/utils/sharedDataTransformers/squashTopStories/index.test.js
